+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added missing monitoring options in the Helm chart values.yaml
+- Support sharding report metrics across multiple instances of the exporter.
+- Add missing monitoring options in the Helm chart values.yaml.
 
 ## [0.4.1] - 2022-04-26